refactor(mempool): move a flow test to the appropriate file
