+++ crates/mempool/src/mempool_test.rs
@@ -763,35 +763,3 @@ fn test_account_nonce_might_decrease_in_commit_block() {
         MempoolContentBuilder::new().with_account_nonces([("0x0", 1)]).build();
     expected_mempool_content.assert_eq(&mempool);
 }
-
-// Flow tests.
-
-#[rstest]
-fn test_flow_commit_block_fills_nonce_gap(mut mempool: Mempool) {
-    // Setup.
-    let tx_nonce_3_account_nonce_3 =
-        add_tx_input!(tx_hash: 1, sender_address: "0x0", tx_nonce: 3, account_nonce: 3);
-    let tx_nonce_5_account_nonce_3 =
-        add_tx_input!(tx_hash: 2, sender_address: "0x0", tx_nonce: 5, account_nonce: 3);
-
-    // Test.
-    for input in [&tx_nonce_3_account_nonce_3, &tx_nonce_5_account_nonce_3] {
-        add_tx(&mut mempool, input);
-    }
-
-    get_txs_and_assert_expected(&mut mempool, 2, &[tx_nonce_3_account_nonce_3.tx]);
-
-    let nonces = [("0x0", 4)];
-    commit_block(&mut mempool, nonces);
-
-    // Assert: hole was indeed closed.
-    let tx_nonce_4_account_nonce_4 =
-        add_tx_input!(tx_hash: 3, sender_address: "0x0", tx_nonce: 4, account_nonce: 4);
-    add_tx_expect_error(
-        &mut mempool,
-        &tx_nonce_4_account_nonce_4,
-        MempoolError::DuplicateNonce { address: contract_address!("0x0"), nonce: nonce!(4) },
-    );
-
-    get_txs_and_assert_expected(&mut mempool, 2, &[tx_nonce_5_account_nonce_3.tx]);
-}
